requirements.txt
fastapi[all]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]